+++ apps/shell/src/components/WorkspaceSwitcher/index.ts
@@ -1,0 +1,3 @@
+export * from './WorkspaceSwitcher.component';
+export * from './WorkspaceSwitcher.types';
+export * from './WorkspaceSwitcher.hook';